+++ RELEASE_NOTES.md
@@ -2,7 +2,8 @@
 
 ### TBD
 
-* Fix issue where the `failedValidation` property was not set for `Content-Type` request validation errors _(PR 420)_
+* Added support for nested controllers _(PR #422, Issue #283)_
+* Fix issue where the `failedValidation` property was not set for `Content-Type` request validation errors _(PR #420)_
 
 ### 0.10.1 (2016-02-24)