+++ RELEASE_NOTES.md
@@ -6,6 +6,10 @@
 * Fix issue where the `failedValidation` property was not set for `Content-Type` request validation errors _(PR #420)_
 * Fix issue where `array` `body` parameters would coerce non-array types into an array _(Issue #438)
 
+### 0.10.2 (2017-05-24)
+
+* Fix issue where swagger-ui shows an error icon trying to validate a swagger doc for an internal API *(Issue 323, 370)*
+
 ### 0.10.1 (2016-02-24)
 
 * Fix issue with initializing middleware with an invalid Swagger document *(Issue 355)*